--- internal/core/routers/router.go
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2022, AcmeStack
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package routers
-
-import (
-	api2 "github.com/acmestack/envcd/internal/core/routers/api"
-	openservice "github.com/acmestack/envcd/internal/core/service"
-	"github.com/gin-gonic/gin"
-)
-
-// InitRouter initialize routing information
-func InitRouter(op *openservice.OpenService) *gin.Engine {
-	r := gin.New()
-	r.Use(gin.Logger())
-	r.Use(gin.Recovery())
-
-	// user auth
-	r.GET("/auth", api2.GetAuth(op))
-
-	//apiv1 := r.Group("/api/v1")
-
-	//apiv1.GET("/tags", v1.GetTags)
-	//apiv1.Use(jwt.JWT())
-	//{
-	//	apiv1.GET("/tags", v1.GetTags)
-	//}
-
-	return r
-}
